Alter logic to create db in $HOME/.wanikani instead of .h2
Orignally we were creating the db in the .h2 folder

The .wanikani folder can hold everything related to the
project such as the db and things I haven't considered yet, such as log
files.

Its a lot cleaner since we can group data
related to this project into one place.