Create WebApi abstraction to standardize WaniKani API URLs
I created an API interface and added an implementation
for V2 of WaniKanis web api. Currently this is only for the User
endpoint but will be expanded as needed. This abstraction will
make it easier to swap out different API implementations if WaniKani
ever decides to create a new web api in the future.