Change each call to the WaniKani API to be async
Most of the operations we will be doing synching up the application
with WaniKani are optional, and don't require and active internet
connection for the application to function.

There are only a few instances that this is not the case,
such as logging in for the first time.

This feels more right, I'll change functions back if they are not used asynccly.