Add very basic(ugly) login page
Add logic to prompt the user for an API key if not logged in.
If they aren't and the API key is valid, store it in the db
and login as that user from now on and have them land
on the home page(Which is blank currently).

Add some basic unit tests for our backend services. Need to figure
out how to do selenium with JavaFX(Kill me).